requirements.txt
beautifulsoup4==4.10.0
requests==2.26.0
google-cloud-storage==1.42.3
google-cloud-secret-manager==2.7.1